Move toward being able to select block textures using marching squares.
Restructure code to move in a direction where it would be possible to
select the textures for a block based on marching squares. Each cube used
by marching cubes can be decomposed into six faces. Each face has four
vertices, these four vertices can be used to select the appropriate
materials.

+++ GutsyStorm/GSTerrainBuffer.h
@@ -12,7 +12,7 @@
 #import "GSVoxel.h" // for GSVoxelBitwiseOp
 
 
-typedef uint16_t GSTerrainBufferElement;
+typedef uint32_t GSTerrainBufferElement;
 
 
 static inline size_t BUFFER_SIZE_IN_BYTES(vector_long3 dimensions)
